Examples for chapter 1 (the Visitor Pattern)
This example is largely identical to the code in the comic.

One note:
All the classes are included in one file for clarity-
I considered creating different physical units for
each logical unit, as one would in a larger project,
but for this - such an division I deemed to be overkill!